using file dialog to load and save

+++ base64.cpp
@@ -53,7 +53,7 @@ std::string base64_encode(BYTE const* buf, unsigned int bufLen) {
 }
 
 std::vector<BYTE> base64_decode(std::string const& encoded_string) {
-  int in_len = encoded_string.size();
+  size_t in_len = encoded_string.size();
   int i = 0;
   int j = 0;
   int in_ = 0;
